fix: move global modals from app.tsx
