+++ src/components/organisms/GlobalModals/index.ts
@@ -1,0 +1,1 @@
+export {default} from './GlobalModals';
